+++ packages/admin/admin/src/theme/componentsTheme/getComponentsTheme.ts
@@ -16,6 +16,7 @@ import { getMuiCheckbox } from "./MuiCheckbox";
 import { getMuiChip } from "./MuiChip";
 import { getMuiDataGrid } from "./MuiDataGrid";
 import { getMuiDateCalendar } from "./MuiDateCalendar";
+import { getMuiDatePicker } from "./MuiDatePicker";
 import { getMuiDialog } from "./MuiDialog";
 import { getMuiDialogActions } from "./MuiDialogActions";
 import { getMuiDialogContent } from "./MuiDialogContent";
@@ -118,5 +119,6 @@ export const getComponentsTheme = (components: Components, theme: Theme): ThemeO
     MuiTypography: getMuiTypography(components.MuiTypography, theme),
     MuiTablePagination: getMuiTablePagination(components.MuiTablePagination, theme),
     MuiDateCalendar: getMuiDateCalendar(components.MuiDateCalendar, theme),
+    MuiDatePicker: getMuiDatePicker(components.MuiDatePicker, theme),
     MuiTableFooter: getMuiTableFooter(components.MuiTableFooter, theme),
 });