+++ demo/admin/src/common/blocks/CallToActionBlock.tsx
@@ -15,12 +15,16 @@ export const CallToActionBlock = createCompositeBlock(
             },
             variant: {
                 block: createCompositeBlockSelectField<CallToActionBlockData["variant"]>({
+<<<<<<< HEAD
                     label: <FormattedMessage id="callToActionBlock.variant" defaultMessage="Variant" />,
                     defaultValue: "Contained",
+=======
+                    defaultValue: "contained",
+>>>>>>> main
                     options: [
-                        { value: "Contained", label: <FormattedMessage id="callToActionBlock.variant.contained" defaultMessage="Contained" /> },
-                        { value: "Outlined", label: <FormattedMessage id="callToActionBlock.variant.outlined" defaultMessage="Outlined" /> },
-                        { value: "Text", label: <FormattedMessage id="callToActionBlock.variant.text" defaultMessage="Text" /> },
+                        { value: "contained", label: <FormattedMessage id="callToActionBlock.variant.contained" defaultMessage="Contained" /> },
+                        { value: "outlined", label: <FormattedMessage id="callToActionBlock.variant.outlined" defaultMessage="Outlined" /> },
+                        { value: "text", label: <FormattedMessage id="callToActionBlock.variant.text" defaultMessage="Text" /> },
                     ],
                 }),
             },