+++ demo/admin/src/products/generator/SelectProductsGrid.cometGen.ts
@@ -11,7 +11,7 @@ export default defineConfig<GQLProduct>({
         { type: "text", name: "title", headerName: "Titel", minWidth: 200, maxWidth: 250 },
         { type: "text", name: "description", headerName: "Description" },
         { type: "number", name: "price", headerName: "Price", maxWidth: 150 },
-        { type: "staticSelect", name: "type", maxWidth: 150, values: [{ value: "Cap", label: "great Cap" }, "Shirt", "Tie"] },
+        { type: "staticSelect", name: "type", maxWidth: 150, values: [{ value: "cap", label: "great Cap" }, "shirt", "tie"] },
         { type: "date", name: "availableSince", width: 140 },
         { type: "dateTime", name: "createdAt", width: 170 },
     ],