+++ packages/admin/admin/CHANGELOG.md
@@ -24,7 +24,7 @@
 
     ```diff
     +import { LocalizationProvider } from "@mui/x-date-pickers";
-    +import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFnsV3";
+    +import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
     +import { enUS } from "date-fns/locale";
 
      // ...