+++ packages/admin/admin/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/admin
 
+<<<<<<< HEAD
 ## 8.0.0-beta.6
 
 ### Major Changes
@@ -432,6 +433,28 @@
 - Updated dependencies [04e308a]
 - Updated dependencies [682a674]
     - @comet/admin-icons@8.0.0-beta.0
+=======
+## 7.25.3
+
+### Patch Changes
+
+-   @comet/admin-icons@7.25.3
+-   @comet/admin-theme@7.25.3
+
+## 7.25.2
+
+### Patch Changes
+
+-   @comet/admin-icons@7.25.2
+-   @comet/admin-theme@7.25.2
+
+## 7.25.1
+
+### Patch Changes
+
+-   @comet/admin-icons@7.25.1
+-   @comet/admin-theme@7.25.1
+>>>>>>> main
 
 ## 7.25.0